+++ release/KoriUI/Locales/esES.lua
@@ -1,8 +1,0 @@
--- Localization file for Spanish/Spain
-local L = LibStub("AceLocale-3.0"):NewLocale("KoriUI", "esES", false, true)
-
-if L then
-    L["AddonName"] = "Kori UI"
-    L["LeftClickOpen"] = "Left Click: Open Installer"
-    L["PlaterPrompt"] = "Insert a Name for the New Plater Profile:"
-end