+++ release/KoriUI/Locales/itIT.lua
@@ -1,8 +1,0 @@
--- Localization file for Italian/Italy
-local L = LibStub("AceLocale-3.0"):NewLocale("KoriUI", "itIT", false, true)
-
-if L then
-    L["AddonName"] = "Kori UI"
-    L["LeftClickOpen"] = "Left Click: Open Installer"
-    L["PlaterPrompt"] = "Insert a Name for the New Plater Profile:"
-end