+++ release/KoriUI/Locales/koKR.lua
@@ -1,8 +1,0 @@
--- Localization file for Korean/Korea
-local L = LibStub("AceLocale-3.0"):NewLocale("KoriUI", "koKR", false, true)
-
-if L then
-    L["AddonName"] = "Kori UI"
-    L["LeftClickOpen"] = "Left Click: Open Installer"
-    L["PlaterPrompt"] = "Insert a Name for the New Plater Profile:"
-end