+++ utils/kori_tooltips.lua
@@ -480,27 +480,8 @@ eventFrame:SetScript("OnEvent", function(self, event, ...)
         C_Timer.After(0.5, function()
             SetupTooltipHook()
 
-            -- Wrap MoneyFrame functions in pcall to suppress Blizzard secret value bug
-            if MoneyFrame_Update then
-                local originalMoneyFrameUpdate = MoneyFrame_Update
-                MoneyFrame_Update = function(...)
-                    pcall(originalMoneyFrameUpdate, ...)
-                end
-            end
-            if SetTooltipMoney then
-                local originalSetTooltipMoney = SetTooltipMoney
-                SetTooltipMoney = function(...)
-                    pcall(originalSetTooltipMoney, ...)
-                end
-            end
-
-            -- Wrap GameTooltip:SetSpellByID in pcall to suppress Blizzard PTRFeedback secret value bug
-            if GameTooltip and GameTooltip.SetSpellByID then
-                local originalSetSpellByID = GameTooltip.SetSpellByID
-                GameTooltip.SetSpellByID = function(...)
-                    pcall(originalSetSpellByID, ...)
-                end
-            end
+            -- Do not replace Blizzard tooltip/money functions; replacing them taints
+            -- tooltip/widget execution paths and can trigger secret value errors.
         end)
     elseif event == "MODIFIER_STATE_CHANGED" then
         OnModifierStateChanged()